converst NotFoundPage and its unit tests to typescript
